split monotone barrier strategy into its own file
